+++ steamclog/src/main/java/com/steamclock/steamclog/Destinations.kt
@@ -1,5 +1,6 @@
 package com.steamclock.steamclog
 
+import android.annotation.SuppressLint
 import android.util.Log
 import timber.log.Timber
 import java.io.File
@@ -41,28 +42,33 @@ internal class SentryDestination : Timber.Tree() {
         val originalMessage = wrapper?.originalMessage ?: message
         val originalThrowable = wrapper?.originalThrowable
 
-        val breadcrumbMessage = generateSimpleLogMessage(
+        // Always add breadcrumb that indicates the log message.
+        Sentry.addBreadcrumb(generateSimpleLogMessage(
             priority,
             includeEmoji = false,
             wrapper,
             message
-        )
+        ), breadcrumbCategory)
 
         when {
             priority == Log.ERROR && originalThrowable != null -> {
-                // If given an original throwable, capture it
-                Sentry.addBreadcrumb(breadcrumbMessage, breadcrumbCategory)
-                Sentry.captureException(originalThrowable)
+                // Check to see if we want to allow or block the Throwable from being reported
+                // as an error.
+                if (SteamcLog.throwableBlocker.shouldBlock(originalThrowable)) {
+                    Sentry.addBreadcrumb("${originalThrowable::class.simpleName} on blocked list, and has " +
+                            "been blocked from being captured as an exception: " +
+                            "${originalThrowable.message}", breadcrumbCategory)
+                } else {
+                    Sentry.captureException(originalThrowable)
+                }
             }
             priority == Log.ERROR -> {
-                // If no throwable given, log error as message (no throwable)
-                Sentry.addBreadcrumb(breadcrumbMessage, breadcrumbCategory)
+                // If no throwable given, capture message as the error
                 Sentry.captureMessage(originalMessage)
             }
             else -> {
-                // Not an error, add breadcrumb only (which should include
-                // both the message and any extra data.
-                Sentry.addBreadcrumb(breadcrumbMessage, breadcrumbCategory)
+                // Not an error; breadcrumb should already be added, so we do not
+                // need to do anything more.
             }
         }
     }
@@ -301,6 +307,7 @@ private fun generateFullLogMessage(priority: Int,
  * Used mostly for Steamclog to report message using Log (not Timber), since using
  * Timber may lead to infinite calls in the library.
  */
+@SuppressLint("LogNotTimber")
 internal fun logToConsole(message: String) {
     Log.v("steamclog", message)
 }